+++ app/src/main/res/values/themes.xml
@@ -23,6 +23,7 @@
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
         <item name="android:windowBackground">@color/main_background</item>
+        <item name="preferenceTheme">@style/AppTheme.PreferenceThemeOverlay</item>
     </style>
 
     <style name="AppTheme.NoActionBar" parent="AppTheme">
@@ -39,4 +40,18 @@
         <item name="windowActionBarOverlay">true</item>
     </style>
 
+    <style name="AppTheme.Preference.SwitchPreference" parent="@style/Preference.SwitchPreference">
+        <item name="android:layout">@layout/preference_material</item>
+        <item name="android:widgetLayout">@layout/preference_widget_switch</item>
+    </style>
+
+    <style name="AppTheme.Preference.SwitchPreferenceCompat" parent="@style/Preference.SwitchPreferenceCompat">
+        <item name="android:layout">@layout/preference_material</item>
+        <item name="android:widgetLayout">@layout/preference_widget_switch</item>
+    </style>
+
+    <style name="AppTheme.PreferenceThemeOverlay" parent="PreferenceThemeOverlay.v14.Material">
+        <item name="switchPreferenceStyle">@style/AppTheme.Preference.SwitchPreference</item>
+        <item name="switchPreferenceCompatStyle">@style/AppTheme.Preference.SwitchPreferenceCompat</item>
+    </style>
 </resources>
