+++ app/src/main/res/xml/pref_general.xml
@@ -14,35 +14,36 @@
   ~ limitations under the License.
   -->
 
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+<android.support.v7.preference.PreferenceScreen
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <PreferenceCategory
+    <android.support.v7.preference.PreferenceCategory
         android:key="basic"
         android:title="@string/basic_setting">
 
-        <SwitchPreference
+        <android.support.v7.preference.SwitchPreferenceCompat
             android:defaultValue="false"
             android:key="safe_data_mode"
             android:summary="@string/auto_load_image_in_wifi_mode"
             android:title="@string/load_image" />
-    </PreferenceCategory>
+    </android.support.v7.preference.PreferenceCategory>
 
-    <PreferenceCategory
+    <android.support.v7.preference.PreferenceCategory
         android:key="about"
         android:title="@string/about">
 
-        <Preference
+        <android.support.v7.preference.Preference
             android:key="help_and_feedback"
             android:title="@string/nav_help" />
 
-        <Preference
+        <android.support.v7.preference.Preference
             android:key="author_email"
             android:summary="@string/info_default"
             android:title="@string/author_email" />
 
-        <Preference
+        <android.support.v7.preference.Preference
             android:key="version_name"
             android:summary="@string/info_default"
             android:title="@string/version_name" />
-    </PreferenceCategory>
-</PreferenceScreen>
+    </android.support.v7.preference.PreferenceCategory>
+</android.support.v7.preference.PreferenceScreen>
